requirements.txt
absl-py==1.3.0
altair==4.2.0
appnope==0.1.3
asttokens==2.2.1
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
blinker==1.5
bqplot==0.12.36
branca==0.6.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colour==0.1.5
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
filelock==3.8.2
flatbuffers==22.11.23
folium==0.13.0
fonttools==4.38.0
gast==0.4.0
gdown==4.6.0
geojson==2.5.0
geojsoncontour==0.4.0
gitdb==4.0.10
GitPython==3.1.29
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
haversine==2.7.0
idna==3.4
importlib-metadata==5.1.0
ipyevents==2.0.1
ipyfilechooser==0.6.0
ipykernel==6.17.1
ipyleaflet==0.17.2
ipython==8.7.0
ipytree==0.2.2
ipywidgets==8.0.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyterlab-widgets==3.0.3
keras==2.11.0
kiwisolver==1.4.4
leafmap==0.14.0
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
OWSLib==0.27.2
packaging==21.3
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
prompt-toolkit==3.0.33
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
pyshp==2.3.1
PySocks==1.7.1
pystac==1.6.1
pystac-client==0.5.1
python-box==6.1.0
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
scipy==1.9.3
scooby==0.7.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.6.2
streamlit==1.15.2
streamlit-folium==0.7.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.6.0
traittypes==0.2.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
wcwidth==0.2.5
Werkzeug==2.2.2
whitebox==2.2.0
whiteboxgui==2.2.0
widgetsnbextension==4.0.3
wrapt==1.14.1
xarray==2022.12.0
xyzservices==2022.9.0
zipp==3.11.0
